+++ src/platform/api/src/modules/user/user.seed.ts
@@ -16,16 +16,22 @@ export async function seedUserModule(database: Kysely<TechMediaDatabase>) {
       is_protected: true,
       name,
       password_hash: hashPassword(password),
-      role: "admin",
+      role: "super-admin",
       status: "active",
       uuid: stable(email)
     })
     .onDuplicateKeyUpdate({
       is_protected: true,
       name,
-      role: "admin",
+      password_hash: hashPassword(password),
+      role: "super-admin",
       status: "active"
     })
+    .execute();
+  await database
+    .updateTable("users")
+    .set({ is_protected: false })
+    .where("id", "!=", 1)
     .execute();
 }